Updates to Jannovar tests.

+++ exomiser-core/src/test/java/de/charite/compbio/exomiser/core/filters/QualityFilterTest.java
@@ -5,6 +5,7 @@
  */
 package de.charite.compbio.exomiser.core.filters;
 
+import de.charite.compbio.exomiser.core.Variant;
 import de.charite.compbio.exomiser.core.filters.VariantFilter;
 import de.charite.compbio.exomiser.core.filters.TargetFilter;
 import de.charite.compbio.exomiser.core.filters.FilterType;
@@ -12,12 +13,14 @@ import de.charite.compbio.exomiser.core.filters.QualityFilter;
 import de.charite.compbio.exomiser.core.filters.FilterResultStatus;
 import de.charite.compbio.exomiser.core.filters.FilterResult;
 import de.charite.compbio.exomiser.core.model.VariantEvaluation;
-import jannovar.exome.Variant;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
+
 import org.junit.Before;
 import org.junit.Test;
+
 import static org.junit.Assert.*;
+
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
@@ -32,9 +35,9 @@ public class QualityFilterTest {
     
     private QualityFilter instance;
 
-    private static final float MIN_QUAL_THRESHOLD = 3.0f;
-    private static final float OVER_THRESHOLD = MIN_QUAL_THRESHOLD + 1.0f;
-    private static final float UNDER_THRESHOLD = MIN_QUAL_THRESHOLD - 1.0f;
+    private static final double MIN_QUAL_THRESHOLD = 3.0f;
+    private static final double OVER_THRESHOLD = MIN_QUAL_THRESHOLD + 1.0f;
+    private static final double UNDER_THRESHOLD = MIN_QUAL_THRESHOLD - 1.0f;
     
     private static VariantEvaluation highQualityPassesFilter;
     private static VariantEvaluation lowQualityFailsFilter;
